+++ NEWS.md
@@ -2,6 +2,12 @@
 
 Bioconductor release 3.12 package.
 
+## Changes in 1.0.5
+
+- `filterIntensity` supports passing of additional parameters to the used
+  filter function (issue
+  [164](https://github.com/rformassspectrometry/Spectra/issues/164)).
+
 ## Changes in 1.0.4
 
 - Fix `show,ProcessingStep` (issue